+++ packages/flutter_elemental/lib/flutter_elemental.dart
@@ -5,5 +5,6 @@ export 'package:flutter_nucleus/flutter_nucleus.dart';
 
 export 'src/build_context.dart';
 export 'src/hooks.dart';
+export 'src/navigator.dart';
 export 'src/nucleus.dart';
 export 'src/storage.dart';